feature: add edit ticket funcionality

+++ src/app/api/tickets/route.ts
@@ -30,6 +30,41 @@ export async function POST(req: Request) {
   }
 }
 
+export async function PUT(req: Request) {
+  const session = await getServerSession(authOptions);
+
+  if (!session || !session.user) {
+    return NextResponse.json({ message: "Not Authorized." }, { status: 401 });
+  }
+
+  const { searchParams } = new URL(req.url);
+
+  const id = searchParams.get("id");
+
+  if (!id) {
+    return NextResponse.json({ message: "Invalid Ticket ID" }, { status: 406 });
+  }
+
+  const data: {
+    name?: string;
+    status?: string;
+  } = await req.json();
+
+  try {
+    await prisma.ticket.update({
+      where: { id: id as string },
+      data: { name: data.name, status: data.status, updated_at: new Date() },
+    });
+    return NextResponse.json({ message: "Success." });
+  } catch (error) {
+    console.error(error);
+    return NextResponse.json(
+      { message: "Couldn't update the ticket." },
+      { status: 500 }
+    );
+  }
+}
+
 export async function DELETE(req: Request) {
   const session = await getServerSession(authOptions);
 
@@ -40,6 +75,10 @@ export async function DELETE(req: Request) {
   const { searchParams } = new URL(req.url);
 
   const id = searchParams.get("id");
+
+  if (!id) {
+    return NextResponse.json({ message: "Invalid Ticket ID" }, { status: 406 });
+  }
 
   try {
     await prisma.ticket.delete({ where: { id: id as string } });
